requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.2
yfinance==0.2.43
yahooquery==2.3.7
vaderSentiment==3.3.2
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.32.3
curl_cffi==0.6.0b9
scipy==1.14.1
transformers==4.44.2
torch>=2.2.0
sentencepiece
safetensors
